Rename base package to arch, and add a module package inside
