+++ src/main/java/com/worldwarofants/game/ConsoleGame.java
@@ -1,5 +1,6 @@
 package com.worldwarofants.game;
 
+import com.worldwarofants.game.arch.module.ModuleManager;
 import com.worldwarofants.game.examplefeature.arch.ExampleModule;
 import com.worldwarofants.game.model.World;
 import com.worldwarofants.game.model.WorldMock;
@@ -11,25 +12,37 @@ import java.util.Scanner;
  *
  * @author Veradux
  */
-class ConsoleGame {
+class ConsoleGame extends ModuleManager {
 
     private Scanner scanner;
+    // TODO make a connection between modules
     private ExampleModule example;
+    private ExampleModule example2;
+    private ExampleModule currentModule;
 
     ConsoleGame() {
         scanner = new Scanner(System.in);
-        // TODO replace WorldMock() with the Actual World that will be used
+        // TODO replace WorldMock() with World(), once architecture is approved.
         World world = new WorldMock();
         example = new ExampleModule(world);
+        example2 = new ExampleModule(world);
+        currentModule = defineStartingModule(world);
+    }
+
+    private ExampleModule defineStartingModule(World world) {
+        return new ExampleModule(world);
     }
 
     public void run() {
+        // This is how you can manually execute commands.
+        currentModule.executeCommand("gameTitle");
         while (true) {
-            example.executeCommand(readInput());
+            // This is how you can let the user execute commands.
+            currentModule.executeCommand(readInput());
         }
     }
 
     private String readInput() {
-        return scanner.nextLine().trim().toLowerCase();
+        return scanner.nextLine().trim();
     }
 }
