+++ src/main/java/com/worldwarofants/game/arch/module/ModuleManager.java
@@ -1,0 +1,4 @@
+package com.worldwarofants.game.arch.module;
+
+public class ModuleManager {
+}
